+++ CHANGELOG.md
@@ -19,8 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Changed
 
-Changes that are not related to specific components
-- [Component] What has been changed
+- [Header] Theme supports max-width and logo-height variables
 
 #### Fixed